Add files via upload
open sketch and run with processing

In this sketch I took the code from my Tutorial 1_2 submission and used the minim library to loop an audio file I created. I was inspired by the films of Stan Brakhage and the music of FourTet so I utilized a Moog Opus 3 Synthesizer to create the sounds. It is a vintage paraphonic synthesizer and I simply held down a simple chord while gradually changing the filter resonance. the result is an uplifting and maybe celestial audiovisual experience. at least I hope. thanks!

The zip folder contains a parent directory, inside of which is the sketch, and the data folder which contains the audio file